move FlowModel class to __init__ file
